remove ChangeLog. wontfix #4
It is already outdated even we want to have it.

As for NEWS or Release Notes, there won't be one as of v0.2.0, at least *not done by me*.

I retract my thought on it in #4 about having a release notes or some sort, not because there shouldn't be one, which I believe there should be.

In fact, I believe clock should have both styles of logs, unfortunately, the direction and the way of movement of clock project don't go smoothly as I'd hope for.

Please reopen #4 if anyone wants to take on this using the style, GNU Coding Standards like, as I mentioned in the issue, complete the logs, and add it back to repository; or create a new issue for NEWS or Release Notes.