remove ifdef for BSD and HOMEBREW (#3)
For BSD, I believe it's done so because at the time of coding, the differences between BSD systems and others are significant. As far as I can see, locale.h should have no problem and all the invoked string functions seem to be from string.h, none of the functions from string.h is used.

The array initializer, since this project conforms to C99, the initialization is supported.

For HOMEBREW, I can not find any related references about this Homebrew algorithm as described in the header comment, so it might just go with BSD.